representing multiple loops and creating and removing them

+++ Triangle Mesh Processor Nov 2014/main/main.pde
@@ -4,7 +4,7 @@
 import processing.opengl.*;                // load OpenGL
 String [] fn=  {"bunny.vts","horse.vts","torus.vts","tet.vts","fandisk.vts","squirrel.vts","venus.vts"};
 int fni=0; int fniMax=fn.length;  
-float weaveTimer = 0;
+int weaveTimer = 0;
 // ** SETUP **
 void setup() { size(800, 800, OPENGL); setColors(); sphereDetail(6); //smooth();
   PFont font = loadFont("Courier-14.vlw"); textFont(font, 12);  // font for writing labels on screen
